pyproject.toml
[project]
name = "nonebot-plugin-today-waifu"
version = "0.1.6"
description = "Nonebot2插件 可以随机抽取群友作为老婆"
authors = [
    {name = "glamorgan9826",email = "glamorgan9826@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.9"
homepage = "https://github.com/glamorgan9826/nonebot-plugin-today-waifu"
repository = "https://github.com/glamorgan9826/nonebot-plugin-today-waifu"

dependencies = [
    "nonebot2[fastapi] (>=2.4.2,<3.0.0)",
    "nonebot-adapter-onebot (>=2.4.6,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "nonebot-plugin-uninfo (>=0.7.3,<0.8.0)",
    "nonebot-plugin-alconna (>=0.57.6,<0.58.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
